+++ x/rollapp/types/rollapp.go
@@ -38,10 +38,12 @@ func NewRollapp(
 }
 
 const (
-	maxAliasLength       = 64
-	maxDescriptionLength = 256
-	maxDataURILength     = 25 * 1024 // 25KB
-	dataURIPattern       = `^data:(?P<mimeType>[\w/]+);base64,(?P<data>[A-Za-z0-9+/=]+)$`
+	maxAliasLength           = 64
+	maxDescriptionLength     = 256
+	maxWebsiteLength         = 256
+	maxGenesisChecksumLength = 64
+	maxDataURILength         = 25 * 1024 // 25KB
+	dataURIPattern           = `^data:(?P<mimeType>[\w/]+);base64,(?P<data>[A-Za-z0-9+/=]+)$`
 )
 
 func (r Rollapp) ValidateBasic() error {
@@ -56,17 +58,19 @@ func (r Rollapp) ValidateBasic() error {
 		return err
 	}
 
-	_, err = sdk.AccAddressFromBech32(r.InitialSequencerAddress)
-	if err != nil {
-		return errorsmod.Wrap(ErrInvalidInitialSequencerAddress, err.Error())
+	if r.InitialSequencerAddress != "" {
+		_, err = sdk.AccAddressFromBech32(r.InitialSequencerAddress)
+		if err != nil {
+			return errorsmod.Wrap(ErrInvalidInitialSequencerAddress, err.Error())
+		}
 	}
 
 	if err = validateBech32Prefix(r.Bech32Prefix); err != nil {
 		return errorsmod.Wrap(ErrInvalidBech32Prefix, err.Error())
 	}
 
-	if r.GenesisChecksum == "" {
-		return errorsmod.Wrap(ErrEmptyGenesisChecksum, "GenesisChecksum")
+	if l := len(r.GenesisChecksum); l == 0 || l > maxGenesisChecksumLength {
+		return errorsmod.Wrap(ErrInvalidGenesisChecksum, "GenesisChecksum")
 	}
 
 	if l := len(r.Alias); l == 0 || l > maxAliasLength {
@@ -102,8 +106,13 @@ func validateMetadata(metadata *RollappMetadata) error {
 		return nil
 	}
 
-	if _, err := url.Parse(metadata.Website); err != nil {
-		return errorsmod.Wrap(ErrInvalidWebsiteURL, err.Error())
+	if metadata.Website != "" {
+		if len(metadata.Website) > maxWebsiteLength {
+			return ErrInvalidWebsiteURL
+		}
+		if _, err := url.Parse(metadata.Website); err != nil {
+			return errorsmod.Wrap(ErrInvalidWebsiteURL, err.Error())
+		}
 	}
 
 	if len(metadata.Description) > maxDescriptionLength {